+++ src/main/java/org/opentripplanner/routing/core/StateData.java
@@ -37,6 +37,8 @@ public class StateData implements Cloneable {
 
     protected boolean everBoarded;
 
+    protected boolean everCarpooled;
+
     protected boolean usingRentedBike;
 
     protected boolean carParked;